requirements.txt
cachetools==5.2.1
grpcio==1.51.1
protobuf==4.21.12
pydantic==1.10.4
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b45
typing_extensions==4.4.0
